requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
drf-yasg>=1.21.6
PyYAML>=6.0
psycopg2-binary>=2.9.6   # optional; needed if you use Postgres
